Generalized routing over platforms and networks.
