+++ build.sbt
@@ -1,8 +1,8 @@
 name := "Conseil"
 version := "0.0.1"
-scalaVersion := "2.12.3"
+scalaVersion := "2.12.4"
 
-val akkaHttpVersion = "10.0.10"
+val akkaHttpVersion = "10.0.11"
 
 libraryDependencies  ++=  Seq(
   "ch.qos.logback" % "logback-classic" % "1.2.3",
