+++ src/main/resources/application.conf
@@ -3,9 +3,13 @@ conseil: {
   port: 1337
 }
 
-tezos: {
-  node: {
-    hostname: "138.197.114.145",
-    port: 8732
+platforms: {
+  tezos: {
+    alphanet : {
+      node: {
+        hostname: "138.197.114.145",
+        port: 8732
+      }
+    }
   }
 }
